extract file_attributes.py
