+++ src/bfabric_scripts/bfabric_save_importresource_sample.py
@@ -15,17 +15,16 @@ History:
 
 from __future__ import annotations
 
-import hashlib
 import json
 import os
 import re
 import sys
 import time
-from pathlib import Path
 
 from loguru import logger
 
 from bfabric import Bfabric, BfabricClientConfig
+from bfabric_scripts.feeder.file_attributes import get_file_attributes
 
 BFABRIC_STORAGE_ID = 2
 
@@ -54,28 +53,6 @@ def save_importresource(client: Bfabric, line: str) -> None:
     logger.info(obj)
     res = client.save(endpoint="importresource", obj=obj)
     print(json.dumps(res, indent=2))
-
-
-def get_file_attributes(file_name_or_attributes: str) -> tuple[str, int, int, str]:
-    """Returns (md5sum, timestamp, size, path) from a given input line, which can either contain
-    the precomputed values (separated by ";") or just a filename."""
-    values = file_name_or_attributes.split(";")
-    if len(values) == 4:
-        return values[0], int(values[1]), int(values[2]), values[3]
-    elif len(values) == 1:
-        filename = values[0].strip()
-        file_path = Path("/srv/www/htdocs") / filename
-        file_stat = file_path.stat()
-        file_size = file_stat.st_size
-        file_unix_timestamp = int(file_stat.st_mtime)
-        hash = hashlib.md5()
-        with file_path.open("rb") as f:
-            for chunk in iter(lambda: f.read(4096), b""):
-                hash.update(chunk)
-        md5_checksum = hash.hexdigest()
-        return md5_checksum, file_unix_timestamp, file_size, filename
-    else:
-        raise ValueError("Invalid input line format")
 
 
 def create_importresource_dict(
